fix: TemplateDoesNotExist when using {% extends %} on main template and two components with same parent template
