+++ src/django_components/context.py
@@ -22,6 +22,10 @@ def make_isolated_context_copy(context: Context) -> Context:
     context_copy = context.new()
     copy_forloop_context(context, context_copy)
 
+    # Required for compatibility with Django's {% extends %} tag
+    # See https://github.com/EmilStenstrom/django-components/pull/859
+    context_copy.render_context = context.render_context
+
     # Pass through our internal keys
     context_copy[_REGISTRY_CONTEXT_KEY] = context.get(_REGISTRY_CONTEXT_KEY, None)
     if _ROOT_CTX_CONTEXT_KEY in context:
