+++ tests/templates/included.html
@@ -1,0 +1,4 @@
+{% extends "simple_template.html" %}
+{% block before_custom %}
+    <div>BLOCK OVERRIDEN</div>
+{% endblock %}
